Efficiency
In a high level interpreted language (regardless which one) built in
functions and language constructs are almost always faster than writing
out the same in code. They work on a finished lower level while the
other has to be interpreted and built from the higher level language.
Thus existing functions are almost always faster even if there are
several of them used.